Fix base address for PE/COFF Windows DLL
When libbacktrace is used directly in a DLL (for example backtrace_create_state is being called DllMain), we can't use GetModuleHandle(NULL) since that returns the module handle of the calling process (executable) and so all our lookups are incorrect. 

Instead, we can use the same function, but pass in the filepath that we already have. This fixes the DLL issue and still works for EXEs.