Don't add DLL notifications if the module is a DLL
This fixes and issue when libbacktrace is used inside a DLL - where
libbacktrace is initialised in DllMain, for example.

If the module is a DLL itself, we shouldn't register for DLL
notifications - that is the job of the executing processing, not of one
of it's modules. Prior to this patch, there can be crashes when using
libbacktrace directly from DLL due to notifications being sent to
invalid addresses.

GetModuleHandle(NULL) returns the handle of the process - so we can use
that alongside our already existing module handle to determine if we are
running as a DLL or EXE.